refactor(api): move route and controller to evaluation user context
