+++ api/lib/application/users/index.js
@@ -191,31 +191,6 @@ const register = async function (server) {
     },
     {
       method: 'PATCH',
-      path: '/api/users/{id}/remember-user-has-seen-assessment-instructions',
-      config: {
-        pre: [
-          {
-            method: securityPreHandlers.checkRequestedUserIsAuthenticatedUser,
-            assign: 'requestedUserIsAuthenticatedUser',
-          },
-        ],
-        validate: {
-          params: Joi.object({
-            id: identifiersType.userId,
-          }),
-        },
-        handler: userController.rememberUserHasSeenAssessmentInstructions,
-        notes: [
-          '- **Cette route est restreinte aux utilisateurs authentifiés**\n' +
-            "- Sauvegarde le fait que l'utilisateur ait vu le didacticiel" +
-            '- L’id demandé doit correspondre à celui de l’utilisateur authentifié',
-          "- Le contenu de la requête n'est pas pris en compte.",
-        ],
-        tags: ['api', 'user'],
-      },
-    },
-    {
-      method: 'PATCH',
       path: '/api/users/{id}/has-seen-challenge-tooltip/{challengeType}',
       config: {
         pre: [
