+++ api/src/evaluation/application/users/user-controller.js
@@ -1,5 +1,13 @@
+import { usecases as libUsecases } from '../../../../lib/domain/usecases/index.js';
 import * as userSerializer from '../../../shared/infrastructure/serializers/jsonapi/user-serializer.js';
 import { evaluationUsecases as usecases } from '../../domain/usecases/index.js';
+
+const rememberUserHasSeenLevelSevenInfo = async function (request, h, dependencies = { userSerializer }) {
+  const authenticatedUserId = request.auth.credentials.userId;
+
+  const updatedUser = await usecases.rememberUserHasSeenLevelSevenInfo({ userId: authenticatedUserId });
+  return dependencies.userSerializer.serialize(updatedUser);
+};
 
 const rememberUserHasSeenNewDashboardInfo = async function (request, h, dependencies = { userSerializer }) {
   const authenticatedUserId = request.auth.credentials.userId;
@@ -8,6 +16,17 @@ const rememberUserHasSeenNewDashboardInfo = async function (request, h, dependen
   return dependencies.userSerializer.serialize(updatedUser);
 };
 
-const userController = { rememberUserHasSeenNewDashboardInfo };
+const rememberUserHasSeenAssessmentInstructions = async function (request, h, dependencies = { userSerializer }) {
+  const authenticatedUserId = request.auth.credentials.userId;
+
+  const updatedUser = await libUsecases.rememberUserHasSeenAssessmentInstructions({ userId: authenticatedUserId });
+  return dependencies.userSerializer.serialize(updatedUser);
+};
+
+const userController = {
+  rememberUserHasSeenLevelSevenInfo,
+  rememberUserHasSeenNewDashboardInfo,
+  rememberUserHasSeenAssessmentInstructions,
+};
 
 export { userController };
