+++ api/tests/evaluation/unit/application/users/user-controller_test.js
@@ -1,3 +1,4 @@
+import { usecases as libUsecases } from '../../../../../lib/domain/usecases/index.js';
 import { userController } from '../../../../../src/evaluation/application/users/user-controller.js';
 import { evaluationUsecases } from '../../../../../src/evaluation/domain/usecases/index.js';
 import { expect, hFake, sinon } from '../../../../test-helper.js';
@@ -24,6 +25,38 @@ describe('Unit | Controller | user-controller', function () {
         hFake,
         { userSerializer },
       );
+
+      // then
+      expect(response).to.be.equal('ok');
+    });
+  });
+
+  describe('#rememberUserHasSeenAssessmentInstructions', function () {
+    let request;
+    let userSerializer;
+    const userId = 1;
+
+    beforeEach(function () {
+      request = {
+        auth: { credentials: { userId } },
+        params: { id: userId },
+      };
+      userSerializer = {
+        serialize: sinon.stub(),
+      };
+
+      sinon.stub(libUsecases, 'rememberUserHasSeenAssessmentInstructions');
+    });
+
+    it('should remember user has seen assessment instructions', async function () {
+      // given
+      libUsecases.rememberUserHasSeenAssessmentInstructions.withArgs({ userId }).resolves({});
+      userSerializer.serialize.withArgs({}).returns('ok');
+
+      // when
+      const response = await userController.rememberUserHasSeenAssessmentInstructions(request, hFake, {
+        userSerializer,
+      });
 
       // then
       expect(response).to.be.equal('ok');
